Dataset sparse copy benchmark
